moving functions to managed_services.py
